ObjC ADM: added RTCAudioDevice protocol [2/N]
Bug: webrtc:14193

+++ modules/remote_bitrate_estimator/packet_arrival_map_test.cc
@@ -102,7 +102,8 @@ TEST(PacketArrivalMapTest, GrowsBufferAndRemoveOld) {
 
   EXPECT_EQ(map.begin_sequence_number(), 43);
   EXPECT_EQ(map.end_sequence_number(), kLargeSeq + 1);
-  EXPECT_EQ(map.end_sequence_number() - map.begin_sequence_number(),
+  EXPECT_EQ(static_cast<size_t>(map.end_sequence_number() -
+                                map.begin_sequence_number()),
             PacketArrivalTimeMap::kMaxNumberOfPackets);
 
   EXPECT_FALSE(map.has_received(41));
